+++ pulse2percept/utils/__init__.py
@@ -18,7 +18,7 @@
 """
 from .base import (PrettyPrint, FreezeError, Frozen, Parametrized, Data,
                    bijective26_name, cached, gamma)
-from .animation import HTMLAnimation
+from .animation import HTMLAnimation, frame_interval
 from .geometry import (cart2pol, pol2cart, delta_angle)
 from .array import is_strictly_increasing, radial_mask, sample, unique
 from .images import center_image, scale_image, shift_image, trim_image
@@ -42,6 +42,7 @@ __all__ = [
     'delta_angle',
     'deprecate_parameter',
     'deprecated',
+    'frame_interval',
     'FreezeError',
     'Frozen',
     'gamma',